Size optimisations

+++ wled00/colors.cpp
@@ -9,7 +9,7 @@
  */
 uint32_t color_blend(uint32_t color1, uint32_t color2, uint16_t blend, bool b16) {
   if(blend == 0)   return color1;
-  uint16_t blendmax = b16 ? 0xFFFF : 0xFF;
+  unsigned blendmax = b16 ? 0xFFFF : 0xFF;
   if(blend == blendmax) return color2;
   uint8_t shift = b16 ? 16 : 8;
 
@@ -52,7 +52,7 @@ uint32_t color_add(uint32_t c1, uint32_t c2, bool fast)
     uint32_t g = G(c1) + G(c2);
     uint32_t b = B(c1) + B(c2);
     uint32_t w = W(c1) + W(c2);
-    uint16_t max = r;
+    unsigned max = r;
     if (g > max) max = g;
     if (b > max) max = b;
     if (w > max) max = w;
